+++ core/src/com/unciv/ui/screens/mapeditorscreen/MapGeneratorSteps.kt
@@ -47,5 +47,4 @@ enum class MapGeneratorSteps(
     Rivers("Let the rivers flow"),
     Resources("Spread Resources", MapGeneratorStepsHelpers.applyResources),
     AncientRuins("Create ancient ruins"),
-    Symmetry("Apply rotational symmetry"),
 }